+++ src/pages/Home.tsx
@@ -1,4 +1,4 @@
-import usePosts from '../hooks/usePost'
+import usePosts from '../hooks/usePosts'
 import classes from './Home.module.css'
 import Post from '../components/Post'
