+++ src/pages/PostDetail.module.css
@@ -1,0 +1,8 @@
+.container {
+  width: 800px;
+  height: auto;
+  text-align: center;
+  display: flex;
+  flex-direction: column;
+  gap: 30px;
+}
